refactor: Update import paths for components in score-card page
